Floating point bits polyfill implementations without js.Math functions.
When encoding, to compute the exponent, we perform a binary search
inside an array of all exact powers of 2. The loop executes
exactly 11 times for Doubles, and 8 times for Floats. It replaces
a call to `log`, `floor` and `pow`, in addition to two conditions
to fix up `e` and `significand`.

When decoding, we also use the same array to get the power of 2
via direct indexing, rather than calling `pow`.