Use custom `ArrayOps` typeclasses and `Comparator` in `j.u.Arrays`.
They replace the previous reliance on Scala's `ClassTag`s and
`Ordering`s.

`ArrayOps` provides generic array operations that do not rely on
Scala's run-time support for unconstrained arrays (which erase to
`Object`). `ArrayCreateOps` provides the ability to create a new
array (previously provided by `ClassTag`s).

Instead of `Ordering`, we directly use `j.u.Comparator`.

Where previously we used Scala specialization, we now rely on
inlining. Effective specialization is achieved because, after
inlining, all the instances of the above typeclasses are
statically known, and they are either `object`s or
`@inline class`es. Not relying on specialization means that we do
not need a separate implementation for reference types (Scala's
specialization never specializes for `AnyRef`).

The only suite of methods for which we cannot inline everything
is `sort`, due to recursion. For those methods, the new
implementation is less efficient, as it will use polymorphic
dispatch to the methods of `ArrayOps` and `Comparator`. There is
however less code duplication in the generated code.

For code size reasons, the `ArrayOps` and `ArrayCreateOps`
instances for primitive types are implemented by the same object,
which also implements `Comparator` for their natural ordering. In
a perfect design, we would have separate objects implementing
`Comparator`, but that results in more generated code, for no
good reason.