Fix the signature of CopyOnWriteArrayList.toArray(T[]).
`Array[T]` erases to `Object` when `T` is unconstrained, but the
Java signature must refer to `Object[]`. Constraining `T` to be
`T <: AnyRef` fixes that.

Remarkably, the method did work before the change, because it
implements an inherited method with the correct signature. The
Scala type system admits this, and Erasure adds a bridge with the
correct signature. However, this made the body less efficient, as
it had to use generic array operations.