IR cleaner: rewrite some intrinsic implicit conversions as no-ops.
There are a number of implicit conversions in the Scala.js standard
library that basically amount to no-ops at run-time. Previously, we
could not directly use them from the javalanglib, because they
would not pass the IR cleaner. This caused a mix of
re-implementation and forced casts at call sites.

We improve the IR cleaner to identify calls to those "intrinsics",
and directly rewrite them as no-ops. This allows to simplify and/or
make safer a number of sites in the javalanglib.

In addition to no-op conversions, we also handle the `truthValue`
conversion, so that we complete the set of operations provided by
`js.DynamicImplicits`.