IR cleaner: erase Scala FunctionN's to JavaScript closures.
`new AnonFunctionN(closure)` is replaced by `closure`, while
`someFunctionN(args)` is replaced by `someFunctionN(args)`. All
function types are rewritten to `any`, and all other references to
the function classes are rewritten to `java.lang.Object`.

This requires that we compile with `-no-specialization` in order
not to get calls to the specialized `apply` methods.

These changes allow to use regular Scala functions, with the arrow
notation, in the javalanglib, including by-name parameters.