Fix initialization warnings
When compiled with -Ysafe-init in Scala 3, the original code receives a warning and
blocks boostrapping with -Ysafe-init.

The reason for the warning is that the types (e.g. NoType) passed to PrimRef is not
fully initialized --- calling `tpe.primRef` would leads to null errors.

While the init checker supports passing an object under initialization as constructor
argument, pattern matching on such an object is not supported.

The refactoring makes the two pattern matches unnecessary.