Optimize `getOrElse{,Update}` for `js.Dictionary` and `js.Map`.
The default implementations go through `get`, which allocates a
`Some` instance.

This improvement causes `ju.regex.PatternCompiler` not to reach
`Some`, `None` and `Option` anymore, which dramatically reduces
its test size, although that improvement will not be seen in
actual codebases that reach `Option` types anywhere else.

This is a forward binary incompatible change, because the overrides
in `js.WrappedDictionary` have a more precise binary signature
(taking a `String` for the `key` instead of an `Object` in the
erased superclass).