Fix #4688: Introduce IR nodes to wrap/unwrap JavaScript exceptions.
This allows to defer to linking the reference to
`scala.scalajs.js.JavaScriptException`. For compatibility reasons,
we must preserve that very class when it is available on the
classpath.

However, if it is not already on the classpath, the emitter
provides an implementation through the linker-private-library.
That implementation does not depend on the scalalib, as checked by
the IR cleaner. We add a linker test in `RunTest` to make sure
that we can indeed correctly link a codebase containing a
`WrapAsThrowable` node, but not any implementation of
`JavaScriptException`.

In a hypothetical Scala.js 2.x, we would move `JavaScriptException`
to the `java.lang` package.