Add JSConstructorDef to replace JSMethodDef(StrLit("constructor")).
Previously, the unique JS constructor of a JS class (or module
class) was represented in the IR as a `JSMethodDef` whose `name`
was a constant `StringLiterl("constructor")`. We had done that
because it superficially looks similar to how the constructor is
actually defined in JavaScript source code.

However, JS constructors are treated in very distinct ways by most
of the linker:

* they are singled out to be compiled as the class definition, and
* their body structure is taken apart to deal with the
  `JSSuperConstructorCall` and its effects on the environment.

This effectively meant that the linker was treating JS constructor
defs as a separate data type with a separate structure already, but
in an inconvenient and unsafe way.

Moreover, it fundamentally prevents to declare a *regular* method
called `constructor` in a JS class, at the IR level. Despite what
one might think, it is actually possible to declare such a method
in a JavaScript class. The two following examples define actual
constructors:

  constructor(...) { ... }
  "constructor"(...) { ... }

but a third variant defines a regular method called `constructor`:

  ["constructor"](...) { ... }

Therefore, it does not make sense for the IR to have a built-in
limitation for `JSMethodDef`s named `"constructor"`.

---

In this commit, we introduce a separate `MemberDef` called
`JSConstructorDef`. Its `body` is structured as a
`JSConstructorBody`, which explicitly separates the statements
coming before and after the super constructor call.

This change simplifies all parts of the linker dealing with JS
constructors, and makes them safer in the process.

We use a deserialization hack to convert a `JSMethodDef` with a
`StringLiteral("constructor")` name into a `JSConstructorDef`.
In this commit, we do not change the compiler back-end yet, to
test the deserialization hack.