Merge the javalanglib into the javalib.
They were separated in fb590e6f8b710186ec2cd4e1b58af09daf7bb734
because scalac could not handle compiling them together (for some
obscure reason related to hijacked classes).

Later, we introduced the IR cleaner on the javalanglib, and so
they had to be kept separate as it did not apply to the javalib.

Now that IR cleaner is also applied to the javalib, we can merge
them back together. Apparently, all the versions of scalac that we
still support can handle it.

The merge prompts 3 kinds of changes in the source code:

* Using Any instead of Object as parameter of some methods that
  are called with unconstrained `T` (this is common practice, and
  corresponds to the fact that scalac interprets `j.l.Object` in
  Java signatures as `Any`.
* Butcher the source signature of `Collections.{min,max}` a bit
  more, because `j.l.Comparable` is now considered a subtype of
  `AnyRef`, so Scala's erasure would still choose `Comparable`
  over `Object`, but Java's erasure requires `Object`.
* Calls to varargs methods now appear to take `Array`s instead of
  varargs.