Use String.repeat from PatternSyntaxException.getMessage.
This was left as a TODO before. Since `repeat` is a JDK 11 method,
and `PatternSyntaxException` was in javalib, it previously could
not see the `def repeat` that was in javalanglib.

Now everything is in the same project, so we can access it.

The implementation in `PatternSyntaxException` had an optimization
for ES 2015 that `String.repeat` did not have. We add that
optimization in `String.repeat` in this commit as well.