Error if Scala.js flags are not in `scalacOptions`
While any Scala.js-focused project would pick up on a warning during local development, unfortunately for many projects/maintainers Scala.js is a cross-build that primarily depends on CI to be tested. If it does not error fatally the problem is unlikely to be noticed.

Context: https://github.com/typelevel/sbt-tpolecat/issues/102, a popular plugin used in many cross-building projects.