In methods of hijacked classes, `this` can be typed as the primitive.
Within an instance method, we know that `this` is non-null. For
hijacked classes, that means we can type it as the corresponding
primitive type.

This improves the generated code for hijacked class methods, as
they do not need to unbox their `$thiz` parameter anymore. This has
a particular importance for `j.l.Character` methods, as they now
take a primitive number instead of a boxed instance. This change of
"private ABI" causes further changes in the codegen of dispatchers
and of direct calls to those methods.

This is a backward binary compatible change because the IR checker
only requires that the type of `This()` nodes be a supertype of the
expected one.