Do not rely on undefined behavior of `substring` in the javalib.
Knowing how `String.substring` was implemented, we have been
abusing it within the javalib with case where the `end` index was
out of bounds.

As we are going to make it optionally compliant, we cannot rely on
that anymore. Instead, we directly use `jsSubstring`, or explicitly
check the bounds ahead of time.