Introduce a CheckedBehavior for StringIndexOutOfBoundsException.
Fundamentally, the primitive operation that causes language-
mandated `StringIndexOutOfBoundsException`s is `String.charAt`.
In order for the linker to recognize it and conditionally apply
checked behavior semantics to it, we introduce a new IR
`BinaryOp.String_charAt` for that operation.

We use a compiler back-end hack to inject that IR node as the body
of `String.charAt`. Using a dedicated primitive method in
`scalajs.runtime`, like we did for `identityHashCode`, would not
work so well here, as we want the binary operator to manipulate a
*primitive* `StringType`, not a `ClassType(BoxedStringClass)`.

For derived methods that are implemented on top of JS primitives,
namely `codePointAt` and `substring`, we use a trick similar to
what we did in `j.u.Arrays`: we introduce carefully-chosen no-op
calls to `charAt`, with the only purpose of checking bounds, before
delegating to the JS methods. Since `s.length()` is not known to be
pure by the optimizer, we use clever tricks not to call it in the
bounds checks of `substring`.