Fix the generation of private "symbols" in the ES 5.1 polyfills.
They were generating fractional values (in
`PrivateFieldsSymbolHolder`) and negative values (in the
`CoreJSLib` polyfill for `PrivateSymbolBuiltin`), which, when
converted to strings, would not follow the expected scheme of 1 to
8 hexadecimal digits.

This was not fundamentally broken, as they remained valid property
names, but was inelegant. We fix both by using a `>>> 0` conversion
to an unsigned 32-bit integer.

Moreover, the variant in `PrivateFieldsSymbolHolder` would actually
cause issues with checked `StringIndexOutOfBoundsException`s, since
it used `String.substring` with a bigger-than-expected index. We
take the opportunity to switch it to `jsSubstring` as well, as we
want this polyfill to be as primitive as possible.