Introduce a UnaryOp for String.length.
This is done mostly for consistency with String.charAt. Together
with `String_+`, those two operations are the only primitives
required to implement any manipulation on strings, so they make a
coherent whole.

In practice, this allows the optimizer to reason about that
operation, in terms of constant-folding and purity. With that, we
need not be afraid of using `length()` in bounds checks, which
allows them to be clearer and more efficient.