Checked behavior for ArrayStoreException.
The fundamental operation that can throw `ArrayStoreException`s is
`Assign(ArraySelect(...), ...)` for a reference type array. We
adapt the emitter to use a `set()` method for reference array types
when either array index out of bounds or array stores are checked
(or both). Assignments for primitive array types are left
unchanged.

In addition to the language, the library function
`System.arraycopy` throws `ArrayStoreException`s in various cases.
The user-space implemention in `System.scala` was already
conforming. We augment the intrinsic version to perform all the
appropriate checks. In many common cases, it is possible to
statically rule out the error cases (when both the `src` and `dest`
arrays are statically known to be primitive arrays) or limit the
error cases to a straightforward compatibility check (when both the
arrays are reference arrays). The full checks are only necessary
when one or both of the arrays are not statically known to be
arrays.

In order to safely show values that cannot be stored, we introduce
a new helper `valueDescription`. It makes sure to report enough
information to debug why a value cannot be stored. Since this is
also useful for debugging cast exceptions, we use the same helper
when throwing `ClassCastException`s, instead of just the `toString`
of the value.