Check ArrayStoreExceptions as part of the arrayErrors checked behavior.
The fundamental operation that can throw `ArrayStoreException`s is
`Assign(ArraySelect(...), ...)` for a reference type array. We
adapt the emitter-generated `set()` method of reference arrays
types to check array store errors in addition to index out of
bounds. The `set()` method of primitive array types is left
unchanged.

In addition to the language, the library function
`System.arraycopy` throws `ArrayStoreException`s in various cases.
The user-space implemention in `System.scala` was already
conforming. We augment the intrinsic version to perform all the
appropriate checks. In many common cases, it is possible to
statically rule out the error cases (when both the `src` and `dest`
arrays are statically known to be primitive arrays) or limit the
error cases to a straightforward compatibility check (when both the
arrays are reference arrays). The full checks are only necessary
when one or both of the arrays are not statically known to be
arrays.

In order to safely show values that cannot be stored in the error
messages, we introduce a new helper `valueDescription`. It makes
sure to report enough information to debug why a value cannot be
stored, while avoiding calling unrestricted `toString()` methods,
which may have side effects and/or cause more errors. Since this is
also useful for debugging cast exceptions, we use the same helper
when throwing `ClassCastException`s, instead of just the `toString`
of the value.