Integrate the JavaScriptException extraction for stack traces in StackTrace.
When we capture the stack trace of a `js.JavaScriptException`, we want
to get the stack trace of the underlying JS error, instead of the point
of instantiation of the `js.javaScriptException`.

Previously, we did this in an override of `fillInStackTrace()`, using
an internal API accessed via reflection. This was not nice, but we had
no choice because `j.l.Throwable` and `j.l.StackTrace` could not unwrap
a `js.JavaScriptException` without accessing Scala.js-specific stuff.

Now, that we have `js.special.unwrapFromThrowable`, we can directly use
that in `StackTrace`. In the process, we simplify the entire logic so
that `StackTrace` doesn't have to manipulate the state of `Throwable`
itself. In turn, this means that no one uses the
`{get,set}StackTraceStateInternal` methods anymore, and therefore we
remove them.