Take the columnNumber in the constructor of `StackTraceElement`.
Historically, we had to use a reflection-based method to set the column
number, because `StackTrace.scala` was compiled separately from
`StackTraceElement.scala`. Unfortunately, that forced
`StackTraceElement` to be mutable whereas it is not supposed to be.

Now, we instead have an additional constructor that directly takes the
column number. We use it in `StackTrace`, and depreate the method
`setColumnNumber`.

The deprecation is symbolic, since that method does not appear in any
binary API. It remains reachable using reflection, which is why we do
not remove it outright. Because of that, `StackTraceElement` is still
effectively mutable, but we have a path to make truly immutable in the
future.