Refactor: Handle constant and non-constant cases in foldBinaryOp.
Previously, we first dispatched between two methods depending on
whether both arguments were literals or not. Although it avoided
some boilerplate, the fact that error cases had to be handled away
from non-error cases made it difficult to reason about correctness.

We now deal with all the constant cases in the same big switch as
everything else.

---

The produced code is unchanged, except for `longLit / 0L` and
`longLit % 0L` where `longLit` is a valid `Int`. The code is now a
bit less efficient as it performs a Long division/remainder,
instead of an Int operation. This is fine because it is an error
case that systematically throws anyway. Reproducing exactly the
same code as before would have required more complicated logic,
which was not worth it.