Embed enough information in the IR for NPE semantics of `throw null`.
In Scala, as in Java, `throw (null: Throwable)` throws a
`NullPointerException`. Since NPEs are Undefined Behaviors in
Scala.js, we previously did not care about this pattern, and
assume it would not happen in the codegen.

If we want to have a checked behavior for NPEs, we need enough
information in the IR for the linker to be able to conditionally
throw accurate `NullPointerException`s in that case.

We do this with a combination of:

* Make `UnwrapFromThrowable(null)` "throw" an NPE, subject to
  undefined behavior (and potentially a checked behavior in the
  future), and
* Emit `Throw(UnwrapFromThrowable(e))` when `e` is nullable, even
  if it cannot be a `JavaScriptException`.

We fix up the non-null-aware IR from previous versions of the
compiler via a deserialization hack.

The produced JavaScript code is basically unchanged after this
commit, since the optimizer can remove the `UnwrapFromThrowable`
in all the cases where the compiler previously omitted them. Some
local variables have different names, and outer pointer checks in
inner JS classes have a worse error path, but that is negligible.