Don't `clean` at every step of the CI script.
Historically, we have `clean`ed the projects between every CI
script because we wanted to force re-linking after changing
settings. This is not necessary anymore, because changing settings
triggers re-linking anyway. Cleaning has the drawback of forcing
re-*compilation* as well, which is a lot of time for the test
suite.

Therefore, we do not `clean` anymore between any two individual
steps. However, we make sure to full-clean everything when we
retry the entire the job.