Fix #4710: Specify our javalib extended binary API.
We introduce a new library `javalibintf`, which is Java-only and
contains no implementation. It only defines the extended binary API
of our javalib.

There were two areas in which the javalib silently exposed an
extended API: TypedArray Buffers, and the column number of
StackTraceElements. We now define well-specified APIs as static
methods for all these secret APIs.

The APIs specified in Java by javalibintf are implemented in
Scala.js by the javalib. They are used from the scalajs-library.
The `library` depends on `javalibintf` in the Provided scope, so
that it is available at compile-time (and in IDEs) but does not
transitively appear in clients of the `library`. This is important
because the APIs are weakly typed when it comes to JavaScript
types such as `TypedArray`s. We do not want regular user to get
access to them too easily.

The `TypedArrayBuffer` API replaces the need for the
`TypedArrayBufferBridge` hack in the build.

The `StackTraceElement` API is not currently used anywhere,
because we did not have any feature in the `library` using them.
So far, only the "reflection" API was specified. We may want to
introduce a safer API in the `library` in the future.

We do publish the `javalibintf`, as it may be used by other
language implementations. It will also be useful to be able to
MiMa-check future versions of the `javalibintf`.