Force the `tpe` of `This` nodes to have the most precise type.
Previously, for `This` nodes, we checked that their `tpe` was a
*subtype* of the expected this type. Now, we force it to always be
the most precise, i.e., the exact this type for the enclosing
class.

This allows to move that check from the IR checker up to the
ClassDef checker, as we do not need the subtyping relationships.

It also allows to remove a hack in the FunctionEmitter where we
had to compute the exact type for `This` nodes for correct boxing
and unboxing of `char` values.

At a fundamental level, this removes the only occurrence of
*subsumption* for terms in the type system of our IR. Every term
must have the most precise type it can have (but can be *used* in
contexts that demand a supertype, obviously).

We introduce a deserialization hack to patch up the `tpe` of `This`
nodes coming from earlier versions of the IR, as the precise type
was not enforced (and was indeed not always as precise as
possible). In this commit, we apply the hack regardless of the IR
version, and do not update the compiler yet, for testing purposes.