Use a single version of the javalib, compiled with Scala 2.12.
Now that the entire javalib is independent of Scala (in the
produced .sjsir files), we do not need to compile it separately for
each version anymore. Instead, we only compile it once, with Scala
2.12, like we do for the linkerPrivateLibrary.

At this point, we still embed the produced .sjsir files in the
scalajs-library, rather than publishing them on their own.

We move the logic related to the minilib to the `javalib` project,
since (by design) it only refers to classes of the `javalib`.