Use only the javalib in the linker tests.
We did not have any test that fundamentally required the full
scalajs-library. Some used through `Predef.println`, but they are
not fundamentally affected by replacing that with
`System.out.println`.

Therefore, we reduce the scope of the "fulllib" previously used by
the linker tests to the "javalib".

This allows to run the LibrarySizeTests in all versions of Scala.

More importantly, as we will separate the javalib in a truly
different jar, keeping the fulllib would have required to load
*two* jars and combine them instead of just one. This reduction in
scope allows us not to increase the setup complexity later on.

The change of `Predef.println` to `System.out.println` does reduce
the surface of IR tested in the backward compat tests and in the
test for no `j.l.Class` in a "hello world" app. This is somewhat
reducing the "value" of these tests.

However, constraining all the tests to the javalib increases the
confidence in the javalib being linkable without the
scalajs-library.