Separate the `javalib` into its own artifact.
Previously, the .sjsir files produced for the `javalib` were
included in the `scalajs-library`. Now, instead, we publish the
`javalib` as its own artifact, and we make the `scalajs-library`
depend on it.

---

The dependency is tricky to set up, because, at *compile*-time, the
`javalib` depends on `library2_12`.

To break the cycle, we decompose the logical `javalib` in two
different contrete projects: `javalibInternal` for *compilation*,
and `javalib` (public) for *packaging* and *publication*.

The `javalibInternal` contains actual sources, and depends on the
`library` via the "no-jar" variant (as before). It is not
published, and nothing dependsOn it.

The public `javalib` has no dependency and is set up like a Java
project (so no dependency on the scalalib either). It contains no
source. However, it takes its `mappings` from the
`javalibInternal`, so its jar contains the products of the
`javalibInternal`. Because if has no dependency, we can safely
have the `library` dependsOn it, which sets up all the correct
dependency information in the published POM files.

---

Having a unique, independent `scalajs-javalib` artifact would allow
other languages to target the Scala.js IR and the `javalib`,
without depending on a particular version of the Scala library.
There are currently no concrete plan like this, but it opens the
door to such a plan.