Fix #4686: Take the `envVars` sbt setting into account.
For `run`, it is read in the scope `project/Config/run`.
For test tasks, it is read in the scope `project/Test` (and not
`project/Test/test`). This is consistent with the JVM, as ensured
by the new scripted test.

We pass it down to the `RunConfig` of `JSEnv`s. If a `JSEnv `does
not support environment variables, and `envVars.value` is
non-empty, the `JSEnv` validator will report an error.