Fix #4726: Use parseFloat(toString()) for BigDecimal.floatValue().
Or rather, a faster version thereof, which does not rescale the
value to follow the public spec of `toString()`.

Similarly, we simplify `doubleValue()` to use `parseDouble()` of
the same `toString()`.

`parseDouble` is provided by the JS engine. We don't have any
evidence either way, but it is not impossible that its native
speed compensates for the conversion to string. `parseFloat`, in
turn, is implemented on top of `parseDouble`, and only has a
fallback for corner cases.

Using `parseDouble` and `parseFloat` for `toString()` is already
what `BigInteger` does. This is one more reason not to bother to
do something cleverer in `BigDecimal`.