Eagerly throw UnsupportedEncodingException in URLDecoder.decode.
Despite what the documentation says, `decode` eagerly throws when
the charset is not supported. This behavior started in JDK 10, when
they added the overload of `decode` taking a `Charset`. With that
addition, maintaining the lazy behavior would probably have
required duplicating the implementation. So in all likelihood, it
will not be coming back. It is still throwing as of JDK 17.