Add an `@JSOperator` annot to explicitly mark JS operator methods.
`@JSOperator` joins the family of "calling convention annotations",
together with `@JSName`, `@JSBracketAccess` and `@JSBracketCall`.
It can only be used on methods whose Scala name corresponds to a
recognized JavaScript operator and have the appropriate signature.

We warn for methods in JS types that have a JavaScript operator
name but have no calling convention annotation. In particular, we
tell the user to add either `@JSName` or `@JSOperator`. In that
case, the method's semantics are backward compatible: they mean
operator semantics if the signature matches, and call semantics if
they don't.

In the future, we will be able to use `@JSOperator` for newly
introduced JavaScript operators, such as `**`, without breaking
backward compatibility.