Fix #4719: Add support for the `**` JavaScript operator.
It is now possible to use `@JSOperator` on a method named `**` in
a JS type. In that case, it will translate to the JavaScript `**`
operator. This is only valid when emitting ES 2016 or later, as it
is not valid ES 2015 code.

Without an explicit `@JSOperator` annotation, such a method still
defaults to being a method call, to preserve backward
compatibility.

We add a definition of `**` on `js.BigInt`. Since bigints are only
supported in ES 2020+ anyway, it is fine to expose it as is.

We also add a definition of `**` on `js.Dynamic`. In that class,
we mark it with an ES 2016 badge.

Note that it is possible to write code using feature detection to
use `js.BigInt`s when they are available, and avoid them otherwise.
When doing that, it is not possible to use `**` unless we require
at least ES 2016.