Expose the linker output *dir* as new tasks {fast,full}LinkJSOutput.
The `fastLinkJSOutput` task runs `fastLinkJS` and extracts the
output directory `File` from the `Attributed[Report]`. Likewise
for `fullLinkJSOutput`.

In the generic implementations within sbt-scalajs, we prefer to
get access to the full attributed `Report`. But in a typical
user build, the knowledge inside the `Report` is superfluous, and
the wrapping gets in the way of writing a simple build file.

The new task keys give a simpler access to what's relevant from a
user point of view. It can also be directly used in a
`print fastLinkJSOutput` command from an orchestrating tool (such
as a JavaScript build tool calling sbt), since it prints as a bare
string with the absolute path.