Checked behavior for NullPointerExceptions.
We introduce a new checked behavior `nullPointers`, which covers
`NullPointerException`s thrown by IR nodes when one of their
arguments is `null`. The relevant nodes are:

* `Apply` and `ApplyStatically` for the receiver,
* `Select` for the qualifier,
* `ArrayLength` and `ArraySelect` for the array,
* `GetClass`, `Clone` and `UnwrapFromException` for their
  respective only arguments.

In the CoreJSLib, we define a new core function `$n(x)`, which
throws when `x === null` and otherwise returns `x`. We give it
such a short name because it gets inserted a lot.

We insert a call to `$n` for all the expressions that must not be
`null` in the nodes above. We avoid doing so when we can prove
that the expression is never `null`. For some simple things, the
emitter can locally decide that (`New`, `This`, `LoadModule` and
literals are some examples).

The optimizer already had null tracking, notably to elide explicit
`null` tests such as those inserted by the compiler for `==`. In
order to communicate the enhanced information to the emitter, we
use a new transient node `AssumeNotNull(x)`. The emitter can then
avoid null checks for those.

We add another transient node `CheckNotNull(x)` for an explicit
null check. This is inserted by the optimizer when it expands a
node that would other need to perform a null check. The canonical
example of that is inlining, which gets rid of the `Apply` node,
but must preserve a null check for the receiver. `CheckNotNull` is
also used internally by the emitter during its translation.

---

When we need to perform a semantics-dependent null check in
user-space, we use the `x.getClass()`-in-statement-position idiom.
This is semantically accurate, since `getClass()` is pure for all
non-null receivers. We make sure that the optimizer can replace it
with a `CheckNotNull` or remove it entirely.

---

Benchmarks suggest that the overhead of these new checks in the
default fastLink configuration are anywhere from 0% to 17%, with
an average of 7%. The code size overhead is of 1.8% in the reversi
demo.

It is possible that smarter optimizations could reduce the overhead
further. For example, after a successful `x.isInstanceOf[T]`, we
could know that `x` is non-null. In fact, we could also know that
it is a `T`. Both pieces of knowledge would be relevant to code
generated for pattern matches. However, our optimizer is currently
not equipped to reason about evaluation-order-dependent knowledge.