Use a semantic-dependent null check for `x.synchronized(...)`.
The compiler itself only generated an explicit NPE for the receiver
of `synchronized`. We convert that to using a `GetClass(x)` in
statement position as a semantics-dependent null check. In theory,
this is a breaking change (and that surfaces in changed tests), as
the NPE was previously reliable and is not anymore. However, it is
very unlikely that anyone would rely on this particular NPE to be
thrown.