Fix #4753: Restrict when GetClass is non-nullable in OptimizerCore.
When its argument is a JS object, `GetClass` actually returns
`null`, so it cannot be considered non-nullable in general.

However, when the type of its argument is neither `AnyType` nor
`j.l.Object` (the latter happens within the code of `Object`
itself), it cannot possibly be a JS object. In that case, we still
consider it non-nullable. The scala-library uses `getClass` on
arrays internally, so this pattern actually shows up.