Isolate sbt plugin tests in a dedicated Jenkins job.
Previously, the sbt plugin tests were bundled with the `tools`
tests for the Scala version used by the sbt plugin. Now, we make
those separate. We run the regular `tools` for all main Scala
versions (on all JDK versions), including that used by sbt. The
sbt plugin tests are isolated in a job that does not depend on the
`scala` version at all. That job also runs `scalastyleCheck`, as
before.