Remove `DoWhile` from the IR; deserialize it to `While` instead.
The `DoWhile` loop was never used by the Scala 3 compiler, and has
not been used in the Scala 2 compiler either since
e1f353a0f8d57411cf9f92aabecd163f72f6c609.

We rewrite it to a `While` loop as a deserialization hack.

We do not remove `DoWhile` from the linker-internal `javascript/`
ASTs at this time. They are now dead code, but could be used by
the `FunctionEmitter` to compile some shapes of `While` loops.