Fix #4784: Wrap all regex Assertions in a group when they are quantified.
Instead of systematically wrapping all possible Assertions in an
additional group--which we used to do for `^` and `$` already--, we
detect Assertions in `compileRepeater` and wrap them there.

This approach is more robust, since there is a single point at
which we need to pay attention to that. It will also emit shorter
output regexes, which may or may not be significant.