Move the uniqueness cache of `Symbol` in a separate file.
This way, we have only one implementation for all the versions of
Scala. This will become more useful as we make its implementation
less trivial.

We also rename it back to `UniquenessCache`, which is the name it
has on the JVM, and reinstate the `Null` lower bound on `V`. There
was no reason to name it differently, and keeping the name may be
better for binary compatibility (even though it is `private[scala]`
and nobody should be using it). It is still different for the key
type, which is fixed to `String`.