Remove the overrides for `s.c.mutable.Buffer`.
They were changed in 757949ebfdb68e32909989f914d8bd106ee59794 to
use a `js.WrappedArray` by default, instead of an `ArrayBuffer`.

Unfortunately, if we want to make the scalalib independent from the
Scala.js library, this is not something we can do anymore.

Ideally, we would override `ArrayBuffer` itself with an
implementation based on `js.Array`. We cannot do that either
because its `array` field is protected, and it is not final.