Apply the IR cleaner to the scalalib and the library-aux.
It is configured differently than for the javalib. We allow
references to `scala.*` (obviously) but disallow references to
`scala.scalajs.*`. We also avoid rewriting Scala functions to JS
functions, since that changes the binary APIs.

We use a local copy of `sjs.runtime.AnonFunctionN` classes in
`scala.runtime.*` to support the codegen of lambdas.