Sort the synthetic methods produced by linking, for stability.
Previously, rerunning the linker twice could produce a different
order for the synthetic methods. This led to unstable outputs.

We make `MethodName` a `Comparable[MethodName]`, so that we can
sort the synthetic methods in a reproducible way. We also make
`TypeRef` a `Comparable[TypeRef]`, which is used in
`MethodName.compareTo`.