Fix the stability of the overload dispatch codegen.
In particular, we always order things by arg count or minimum
arg count.

This ensures that the `case`s of the match are in a stable
order, and that within cases, the alternatives are in a stable
order as well.

Moreover, since we force that order before calling nested
codegen functions, we ensure that inner calls to
`freshLocalIdent`s happen in a stable order as well.