Ensure the stability of top-level and static exports codegen.
Since `groupBy` returns a unordered `Map`, we could potentially
produce unstable code, although I could not observe any issue.

We make sure that the codegen is stable by using an alternative
stable group-by implementation. We already had such an
implementation, although it altered the relative order of the
values within. We fix it in the process.